fix: convert amount to integer before sending to Supabase to resolve invalid bigint syntax error

+++ daily_savings_app/lib/models/savings_entry.dart
@@ -16,7 +16,7 @@ class SavingsEntry {
   Map<String, dynamic> toMap() => {
         'id': id,
         'entry_date': date,
-        'amount': amount,
+        'amount': amount.toInt(),
         'category': category,
         'note': note,
       };
